% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scot.R
\name{get_scot}
\alias{get_scot}
\title{Get data from the Scottish Open Data Platform}
\usage{
get_scot(
  type = c("addresses", "addresstypes", "allowancesgrouptransactions",
    "allowancetransactions", "bills", "billstages", "billstagetypes", "billtypes",
    "committeeroles", "committees", "committeetypelinks", "constituencies",
    "crosspartygrouproles", "crosspartygroups", "departments", "directorates",
    "directoratetypes", "electionreasondates", "electionstatuses", "emailaddresses",
    "emailaddresstypes", "events", "genders", "governmentroles", "legislationdetails",
    "membercrosspartyroles", "memberelectionconstituencystatuses",     
    "memberelectionregionstatuses", "membergovernmentroles", "memberparties",
    "memberpartyroles", "members", "motionsquestionsanswersmotions", "parties",
    "partyroles", "personcommitteeroles", "petitioncountries", "petitiondiscussions",
    "petitiononbehalfofs", "petitions", "petitionsignatures", "petitionstatuses",
    "petitionsubjectcategories", "regions", "registerofinterest", "externalresearches",
    "sessions", "telephones", "telephonetypes", "websites", "websitetypes")
)
}
\arguments{
\item{type}{API to call}
}
\value{
Scottish Open Data
}
\description{
Get data from the Scottish Open Data Platform
}
\examples{
\dontrun{
get_scot(type = "addresses")
}

}
