% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlscot.R
\name{parlscot}
\alias{parlscot}
\title{Download multiple days from the Parliamentary Record}
\usage{
parlscot(start_date = NULL, end_date = NULL)
}
\arguments{
\item{start_date}{From}

\item{end_date}{Until}
}
\value{
Parliamentary Record
}
\description{
Download multiple days from the Parliamentary Record
}
